Make thrift parse/extract and include thrift files in configurations
Add an additional configurations parameter to make thrift
parse jars found there.

If so, make the task depend on the configuration and
parse the jars to find *.thrift files inside.

If found, extract them in a per-jar directory and include
the related directory, by making adding the argument
-I directory
to the thrift command line.

This feature can be activated with a configuration like

compileThrift {
    generator 'java', ''
    configuration 'compile'
}

Change-Id: Ia0d9934b5b767f956370cf21d98242eb77ef9674